+++ src/stories/bookShelf/BookShelf.stories.tsx
@@ -1,15 +1,15 @@
-import Bookshelf from '@/ui/Bookshelf';
+import BookShelf from '@/v1/bookShelf/BookShelf';
 import { Meta, StoryObj } from '@storybook/react';
 
-const meta: Meta<typeof Bookshelf> = {
-  title: 'Bookshelf/Bookshelf',
-  component: Bookshelf,
+const meta: Meta<typeof BookShelf> = {
+  title: 'bookShelf/BookShelf',
+  component: BookShelf,
   tags: ['autodocs'],
 };
 
 export default meta;
 
-type Story = StoryObj<typeof Bookshelf>;
+type Story = StoryObj<typeof BookShelf>;
 
 // TODO: 스토리북에서 이미지를 불러올 때 발생하는 CORS 이슈 해결
 export const Default: Story = {
