+++ src/ui/BookArchive/BookArchiveForUnAuth.tsx
@@ -1,6 +1,6 @@
 import useUnAuthRecommendedBookshelfQuery from '@/queries/recommend/useUnAuthRecommendedBookshelfQuery';
 import { Skeleton, VStack } from '@chakra-ui/react';
-import Bookshelf from '../Bookshelf';
+import Bookshelf from '@/v1/bookShelf/BookShelf';
 
 const BookArchiveForUnAuth = () => {
   const { data, isSuccess, isLoading } = useUnAuthRecommendedBookshelfQuery();
