+++ src/v1/bookGroup/detail/BookGroupInfo.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 
 import Badge from '@/ui/Base/Badge';
 import BookCover from '@/ui/book/BookCover';
-import BookGroupStatus from '@/ui/bookgroup/BookGroupStatus';
+import BookGroupStatus from '@/v1/bookGroup/BookGroupStatus';
 import { IconArrowLeft, IconCalendar, IconMembers } from '@public/icons';
 import { DATA_URL } from '@/constants/dataUrl';
