+++ src/v1/bookGroup/detail/ShortMemberInfo.tsx
